fix: make the tests work
